+++ azure/rs_azure_networking/changelog.md
@@ -1,5 +1,7 @@
-LB Plugin changelog
+Networking Plugin changelog
 
 v1.0
 -----
 - initial release
+- Supports LB Resource
+- Supports Network Interface Resource